Restore combined E2E tests
