--- sycl/test-e2e/WorkGroupMemory/basic_usage_test_ptr.cpp
@@ -1,26 +1,0 @@
-// Pointer-types version of the basic usage test.
-
-// UNSUPPORTED: hip
-// UNSUPPORTED-TRACKER: https://github.com/intel/llvm/issues/17339
-
-// XFAIL: spirv-backend
-// XFAIL-TRACKER: https://github.com/intel/llvm/issues/18230
-
-// RUN: %{build} -o %t.out
-// RUN: %{run} %t.out
-
-#include <sycl/queue.hpp>
-
-sycl::queue q;
-
-#include "./basic_usage_common.hpp"
-
-int main() {
-  test_ptr<float *>();
-  test_ptr<int *>();
-  test_ptr<char *>();
-  test_ptr<uint16_t *>();
-  if (q.get_device().has(sycl::aspect::fp16))
-    test_ptr<sycl::half *>();
-  return 0;
-}
