+++ sycl/test-e2e/bindless_images/vulkan_interop/vulkan_sycl_image_interop_read_2d.cpp
@@ -1,4 +1,11 @@
-// Shared implementation for the Vulkan/SYCL 2D image read interop tests.
+// REQUIRES: aspect-ext_oneapi_bindless_images
+// REQUIRES: aspect-ext_oneapi_external_memory_import || (windows && level_zero && aspect-ext_oneapi_bindless_images)
+// REQUIRES: vulkan
+
+// RUN: %{build} %link-vulkan -o %t.out %if target-spir %{ -Wno-ignored-attributes %}
+
+// UNSUPPORTED: linux
+// UNSUPPORTED-TRACKER: GSD-12357
 
 // XFAIL: windows && gpu-intel-dg2
 // XFAIL-TRACKER: https://github.com/intel/llvm/issues/21985
@@ -7,18 +14,86 @@
 // XFAIL-TRACKER: https://github.com/intel/llvm/issues/21986
 
 /*
-  Run ALL the vulkan formats through the gauntlet. sampled and unsampled.
-  This entire test takes less than 30 seconds on a slow machine.  MUCH faster
-  (and more complete coveraage) than SFINAE based approach.
+    Run ALL the vulkan formats through the gauntlet. sampled and unsampled.
+    This entire test takes less than 30 seconds on a slow machine.  MUCH faster
+   (and more complete coveraage) than SFINAE based approach.
 
-  IF a particular variant is having problems on some platform, please do NOT
-  just disable the whole test, instead use   RUN~IF (SOMETHING) yadda-yadda
-  to enable/disable that variant.
+    IF a particular variant is having problems on some platform, please do NOT
+   just disable the whole test, instead use   RUN~IF (SOMETHING) yadda-yadda
+    to enable/disable that variant.
 
-  For semaphore testing, we run just a sampling. Note, that on Linux if there
-  is a failure in the first section, then likely ALL semaphore tests afterwards
-  will fail. This is being tracked as a separate issue.
+    For semaphore testing, we run just a sampling. Note, that on Linux if there
+   is a failure in the first section, then likely ALL semaphore tests afterwards
+   will fail. This is being tracked as a separate issue.
+
 */
+
+// RUN: %{run} %t.out --type float --channels 1 32x33
+// RUN: %{run} %t.out --type float --channels 2 32x33
+// RUN: %{run} %t.out --type float --channels 4 32x33
+// RUN: %{run} %t.out --type half --channels 1 32x33
+// RUN: %{run} %t.out --type half --channels 2 32x33
+// RUN: %{run} %t.out --type half --channels 4 32x33
+// RUN: %{run} %t.out --type int32 --channels 1 32x33
+// RUN: %{run} %t.out --type int32 --channels 2 32x33
+// RUN: %{run} %t.out --type int32 --channels 4 32x33
+// RUN: %{run} %t.out --type uint32 --channels 1 32x33
+// RUN: %{run} %t.out --type uint32 --channels 2 32x33
+// RUN: %{run} %t.out --type uint32 --channels 4 32x33
+// RUN: %{run} %t.out --type int16 --channels 1 32x33
+// RUN: %{run} %t.out --type int16 --channels 2 32x33
+// RUN: %{run} %t.out --type int16 --channels 4 32x33
+// RUN: %{run} %t.out --type uint16 --channels 1 32x33
+// RUN: %{run} %t.out --type uint16 --channels 2 32x33
+// RUN: %{run} %t.out --type uint16 --channels 4 32x33
+// RUN: %{run} %t.out --type uint8 --channels 1 32x33
+// RUN: %{run} %t.out --type uint8 --channels 2 32x33
+// RUN: %{run} %t.out --type uint8 --channels 4 32x33
+// RUN: %{run} %t.out --type int8 --channels 1 32x33
+// RUN: %{run} %t.out --type int8 --channels 2 32x33
+// RUN: %{run} %t.out --type int8 --channels 4 32x33
+// RUN: %{run} %t.out --type unorm8 --channels 1 32x33
+// RUN: %{run} %t.out --type unorm8 --channels 2 32x33
+// RUN: %{run} %t.out --type unorm8 --channels 4 32x33
+// RUN: %{run} %t.out --type float --channels 1 --sampled 32x33
+// RUN: %{run} %t.out --type float --channels 2 --sampled 32x33
+// RUN: %{run} %t.out --type float --channels 4 --sampled 32x33
+// RUN: %{run} %t.out --type half --channels 1 --sampled 32x33
+// RUN: %{run} %t.out --type half --channels 2 --sampled 32x33
+// RUN: %{run} %t.out --type half --channels 4 --sampled 32x33
+// RUN: %{run} %t.out --type int32 --channels 1 --sampled 32x33
+// RUN: %{run} %t.out --type int32 --channels 2 --sampled 32x33
+// RUN: %{run} %t.out --type int32 --channels 4 --sampled 32x33
+// RUN: %{run} %t.out --type uint32 --channels 1 --sampled 32x33
+// RUN: %{run} %t.out --type uint32 --channels 2 --sampled 32x33
+// RUN: %{run} %t.out --type uint32 --channels 4 --sampled 32x33
+// RUN: %{run} %t.out --type int16 --channels 1 --sampled 32x33
+// RUN: %{run} %t.out --type int16 --channels 2 --sampled 32x33
+// RUN: %{run} %t.out --type int16 --channels 4 --sampled 32x33
+// RUN: %{run} %t.out --type uint16 --channels 1 --sampled 32x33
+// RUN: %{run} %t.out --type uint16 --channels 2 --sampled 32x33
+// RUN: %{run} %t.out --type uint16 --channels 4 --sampled 32x33
+// RUN: %{run} %t.out --type uint8 --channels 1 --sampled 32x33
+// RUN: %{run} %t.out --type uint8 --channels 2 --sampled 32x33
+// RUN: %{run} %t.out --type uint8 --channels 4 --sampled 32x33
+// RUN: %{run} %t.out --type int8 --channels 1 --sampled 32x33
+// RUN: %{run} %t.out --type int8 --channels 2 --sampled 32x33
+// RUN: %{run} %t.out --type int8 --channels 4 --sampled 32x33
+// RUN: %{run} %t.out --type unorm8 --channels 1 --sampled 32x33
+// RUN: %{run} %t.out --type unorm8 --channels 2 --sampled 32x33
+// RUN: %{run} %t.out --type unorm8 --channels 4 --sampled 32x33
+
+// RUN: %{run} %t.out --type float --channels 1 32x33 --semaphores
+// RUN: %{run} %t.out --type float --channels 4 32x33 --semaphores
+// RUN: %{run} %t.out --type half --channels 1 32x33 --semaphores
+// RUN: %{run} %t.out --type uint32 --channels 4 32x33 --semaphores
+// RUN: %{run} %t.out --type uint16 --channels 2 32x33 --semaphores
+// RUN: %{run} %t.out --type int8 --channels 1 32x33 --semaphores
+// RUN: %{run} %t.out --type float --channels 4 --sampled 32x33 --semaphores
+// RUN: %{run} %t.out --type int32 --channels 4 --sampled 32x33 --semaphores
+// RUN: %{run} %t.out --type int16 --channels 4 --sampled 32x33 --semaphores
+// RUN: %{run} %t.out --type uint8 --channels 2 --sampled 32x33 --semaphores
+// RUN: %{run} %t.out --type unorm8 --channels 4 --sampled 32x33 --semaphores
 
 // clang-format off
 /*
@@ -50,8 +125,6 @@
   NOTE: --linear is not currently working with 2D images on Linux
 */
 // clang-format on
-
-#pragma once
 
 #include "vulkan_setup.hpp"
